fix: resolve CI failures — CltvExpiry uint32→uint64, golangci.yml version
- clients/lnd/client.go: InvoiceDefaults.CltvExpiry was uint32 but the
  invoicesrpc.AddHoldInvoiceRequest.CltvExpiry field is uint64; caused
  'cannot use ... as uint64' vet error in CI
- .golangci.yml: remove version: "2" header which conflicted with the
  linters-settings key (v1 format); now unambiguously v1 compatible

+++ clients/lnd/client.go
@@ -35,7 +35,7 @@ type Client struct {
 // InvoiceDefaults holds configurable parameters for hold invoice creation.
 type InvoiceDefaults struct {
 	ExpirySeconds   int64  // How long the invoice is valid (default: 3600)
-	CltvExpiry      uint32 // CLTV delta blocks (default: 40)
+	CltvExpiry      uint64 // CLTV delta blocks (default: 40)
 	FeeRateSatVbyte uint64 // On-chain fee rate in sat/vbyte (default: 1)
 }
